Avoid decreasing acc gain at large errors
The error between the estimated up direction and the upward direction measured by the accelerometer is calculated by a cross product. Both vectors have magnitude 1, so the norm of the error is just the absolute value of sin of the angle between the vectors. For angles with an absolute value of more than 90 degrees the magnitude of the error would decrease, slowing down convergence. This commit introduces a check for those large angles and ensure that the magnitude of the error always increases when the angle increases.